requirements.txt
streamlit
pandas
numpy
scikit-learn
textblob
python-dateutil

